+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.31.0] - 2024-02-09
+
+### Added
+
+* You can now also edit homebrew camping activities
+
+### Fixed
+
+* Prevent homebrew kingdom activities from being overriden
+
 ## [0.30.0] - 2024-02-09
 
 ### Added